requirements.txt
qrcode[pil]>=7.4.2,<8.0.0
pyzbar>=0.1.9,<0.2.0
opencv-python>=4.8.0,<5.0.0
pyautogui>=0.9.54,<0.10.0
pycryptodome>=3.19.0,<4.0.0
pillow>=10.0.0,<11.0.0
configparser>=5.3.0,<6.0.0
numpy>=1.24.0,<2.0.0